requirements.txt
torch
torchvision
matplotlib
numpy
mtcnn
opencv-python
pillow
pandas